style: 182-rename file layout.ts into error.ts
